LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbcursor_advance_index2-expected.txt


PASS IDBCursor.advance() - attempt to pass a count parameter that is not a number
